icon story, break connect field into separate file
